requirements.txt
aiohttp
urllib3
discord.py
tinydb
tabulate